Fixes #20: Make "Enter" key on params go to detail
Default button type is "submit", so "Go" button was already a submit button.
This caused the Enter key to already be submitting the form, which I guess
didn't make any difference because there's no url associated w/the form.

Adding a "(ngSubmit)" event with handler "onSubmit()" in the component allows us
to hardcode a router navigation request, so that was easy enough.